Changes the packager to accept the folder location instead of zip and introduces .bbwpignore.
Implements issue blackberry/BB10-Webworks-Packager#200
It introduces the concept of .bbwpignore to make sure that all the files
are filtered and not bundled with the app.

The spec for .bbwpignore follows that of .gitignore without support for
negations of patterns.
Key points from the spec are -
1. A leading slash matches the beginning of the pathname. For example,
"/*.c" matches "cat-file.c" but not "mozilla-sha1/sha1.c".
2. If the pattern does not contain a slash /, it will be treated as a
global pattern and *.c will match all files in root and subdirectories.
3. Directories must end with /
4. A line starting with # serves as a comment.
5. Blank lines are ignored.
6. ! which negates the pattern is NOT SUPPORTED
 A full spec for .gitignore can be found here -
http://www.kernel.org/pub/software/scm/git/docs/gitignore.html

Wrench has been modified to allow for a filter in the
copyDirSyncRecursive method to allow filtering on files.
Wrench code has been copied from -
https://github.com/ryanmcgrath/wrench-js
and then modified and placed in third_party/wrench folder where all
modifications are clearly marked.

Windows: If you use quotes in your path you provide on windows it must not end with a / because
it causes node to process the to think it is escaping the end " instead
of thinking it as the end of the string.
Take away is that you should use
bbwp "x/y/z" and not bbwp "x/y/z/"
This might get fixed in future version of node.js. Currently tested
against 0.6.10.

Reviewed By: James Keshavarzi <jkeshavarzi@rim.com>
tested By: Tracy Li <tli@rim.com>