Add log level params to bbwp
Fixes Issue # 221

This fix was added mainly for phonegap where a lot of the warnings being
shown were just noise and causing confusion.

A new loglevel argument was added:
--loglevel verbose: the default and the current behavour of bbwp. logs everything
--loglevel warn: logs warning and error messages
--loglevel error: logs error messages

--verbose and -v were depricated

Not providing a loglevel sets the level to verbose.

A new method logger.log was added and some of the more high level
messages were added to it that will always log (regardless of the
loglevel)

Reviewed By: James Keshavarzi <jkeshavarzi@rim.com>